Successfully merged latest changes from master into the OpenTelemetry instrumentation branch.

Resolved conflicts in:
- src/main.rs: Combined metrics initialization with new features (rate limiting, request ID middleware, admin routes)
- src/routes/sessions.rs: Merged tracing instrumentation with refactored endpoints (removed base64 encoding)
- src/services/storage.rs: Combined tracing with cleaner From trait usage
- src/tasks/cleanup.rs: Merged metrics tracking with rate limiter cleanup task

Updated AdminState to include metrics for cleanup endpoint.

All 145 tests passing:
- 96 unit tests
- 7 admin endpoint tests
- 8 API key tests
- 15 metrics tests
- 18 session tests
- 1 migration test

+++ src/services/captcha.rs
@@ -26,11 +26,11 @@ impl CaptchaService {
     pub fn generate(
         &self,
         text: Option<String>,
-        difficulty: i32,
-        width: i32,
-        height: i32,
+        difficulty: i64,
+        width: i64,
+        height: i64,
         dark_mode: bool,
-        compression: i32,
+        compression: i64,
     ) -> Result<(String, Vec<u8>)> {
         let captcha_text = text.unwrap_or_else(|| Self::generate_random_text(5));
